project/05_Advanced_RAG/requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pypdf>=4.2.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
numpy>=1.26.0
plotly>=5.22.0
python-dotenv>=1.0.1
